Move place logic to store array of positions and simplify covers.
