+++ computer_player.rb
@@ -1,0 +1,9 @@
+require './player'
+
+class ComputerPlayer < Player
+
+  def initialize
+    @name = "HAL 9000"
+  end
+
+end
