+++ grid.rb
@@ -1,98 +1,3 @@
-#when you split into separate files, remember to require in each
-class Player
-
-  attr_reader :name
-
-  def initialize(name = "Dave")
-    @name = name
-  end
-
-end
-
-class HumanPlayer < Player
-
-end
-
-class ComputerPlayer < Player
-  def initialize
-    @name = "HAL 9000"
-  end
-end
-
-class Ship
-
-  attr_reader :length
-
-  def initialize(length)
-    #Initializes ship properties and sets size of ship
-    @length = length
-    @placed = false
-    @hit_counter = 0
-  end
-
-  def place(x, y, across)
-    #Places a ship on the board
-    return false if @placed
-    @placed = true
-    @across = across
-    @x = x
-    @y = y
-  end
-
-  def covers?(x, y)
-    #TO DO: Store all the locations that a ship covers in an array
-    #Checks if ship occupies a given space
-    @across ? (@x..@x+@length-1).include?(x) && @y == y : (@y..@y+@length-1).include?(y) && @x == x
-  end
-
-  def overlaps_with?(other_ship)
-    #Checks if two ships overlap
-    if @across
-      (@x..@x+length-1).each do |i|
-        if other_ship.covers?(i, @y)
-          return true
-        end
-      end
-    else
-      (@y..@y+length-1).each do |i|
-        if other_ship.covers?(@x, i)
-          return true
-        end
-      end
-    end
-    false
-  end
-
-  def fire_at(x, y)
-    #Increments hit_counter on successful hits
-    if covers?(x, y)
-      @hit_counter += 1
-      true
-    else
-      false
-    end
-  end
-
-  def sunk?()
-    #Checks if ships has been hit max number of times
-    @length == @hit_counter
-  end
-
-end
-
-class Position
-
-  attr_accessor :occupied, :fired_at
-
-  def initialize()
-    #Sets default properties of Position
-    @occupied = false
-    @fired_at = false
-  end
-
-end
-
-
 class Grid
 
   def initialize()
