+++ human_player.rb
@@ -1,0 +1,9 @@
+require './player'
+
+class HumanPlayer < Player
+
+  def initialize(name = "Dave")
+    @name = name
+  end
+
+end
