+++ player.rb
@@ -1,0 +1,5 @@
+class Player
+
+  attr_reader :name
+
+end
